Use a PushProducer to keep the stream more lively
Instead of directly writing from the TCP client to the Request object, put
a PushProducer in the chain which can respond to
pauseProducing/resumeProducing. This enables us to drop frames when the
network requeset is not able to keep up with it. For now I've also
commented out the MJPEG refreshing on the browser side because it seems to
crash chrome.